Runtime: separate governance host functions and tests into separate mod.
