+++ crates/runtime/src/logic/host_functions.rs
@@ -1,6 +1,8 @@
 
 mod blobs;
 mod storage;
+mod governance;
 
 pub use blobs::*;
 pub use storage::*;
+pub use governance::*;
